cambio

+++ client/src/componentes/services/ServiceList.jsx
@@ -1,17 +1,22 @@
-import React from 'react';
+import React, { useState } from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { FaTrashAlt, FaPen } from 'react-icons/fa';
-import "./ServiceList.css";
+import './ServiceList.css';
 
 function ServiceList({ services, onServiceUpdate, onServiceEdit, userType, onContactProvider }) {
-    const deleteService = async (id) => {
+    const [showModal, setShowModal] = useState(false);
+    const [serviceToDelete, setServiceToDelete] = useState(null);
+
+    const handleDelete = async () => {
+        if (!serviceToDelete) return;
+
         try {
-            const response = await fetch(`http://localhost:5000/api/services/${id}`, {
+            const response = await fetch(`http://localhost:5000/api/services/${serviceToDelete}`, {
                 method: 'DELETE',
                 headers: {
                     'Authorization': `Bearer ${localStorage.getItem('token')}`,
-                    'Content-Type': 'application/json'
-                }
+                    'Content-Type': 'application/json',
+                },
             });
 
             if (!response.ok) {
@@ -20,9 +25,11 @@ function ServiceList({ services, onServiceUpdate, onServiceEdit, userType, onCon
             }
 
             onServiceUpdate();
+            setShowModal(false);
         } catch (error) {
             console.error('Error al eliminar servicio:', error);
             alert('Error al eliminar servicio: ' + error.message);
+            setShowModal(false);
         }
     };
 
@@ -33,23 +40,27 @@ function ServiceList({ services, onServiceUpdate, onServiceEdit, userType, onCon
                     <div key={service._id} className="col-md-4 mb-1">
                         <div className="card h-100 p-3 border border-dark">
                             <div className="row g-0">
-                                <div className="col-md-4 d-flex justify-content-center align-items-start" style={{ padding: 0, marginTop: '20px' }}>
-                                    {service.images && service.images.map((img, index) => (
-                                        <img
-                                            key={index}
-                                            src={`http://localhost:5000/${img}`}
-                                            alt={service.title}
-                                            className="card-img-top img-fluid"
-                                            style={{ objectFit: 'cover', width: '100%', height: '100px' }}
-                                        />
-                                    ))}
+                                <div
+                                    className="col-md-4 d-flex justify-content-center align-items-start"
+                                    style={{ padding: 0, marginTop: '20px' }}
+                                >
+                                    {service.images &&
+                                        service.images.map((img, index) => (
+                                            <img
+                                                key={index}
+                                                src={`http://localhost:5000/${img}`}
+                                                alt={service.title}
+                                                className="card-img-top img-fluid"
+                                                style={{ objectFit: 'cover', width: '100%', height: '100px' }}
+                                            />
+                                        ))}
                                 </div>
                                 <div className="col-md-8">
                                     <div className="card-body">
                                         <h5 className="card-title text-center">{service.title}</h5>
                                         <p className="card-text">{service.description}</p>
                                         <div className="d-flex justify-content-between align-items-center">
-                                        <p className="card-text mb-0 ms-auto text-end">{service.price} Gs</p>
+                                            <p className="card-text mb-0 ms-auto text-end">{service.price} Gs</p>
                                         </div>
                                         {userType === 'provider' && (
                                             <div className="button-group mt-3 d-flex justify-content-end">
@@ -62,9 +73,8 @@ function ServiceList({ services, onServiceUpdate, onServiceEdit, userType, onCon
                                                 <button
                                                     className="btn btn-white"
                                                     onClick={() => {
-                                                        if (window.confirm('¿Estás seguro de eliminar este servicio?')) {
-                                                            deleteService(service._id);
-                                                        }
+                                                        setServiceToDelete(service._id);
+                                                        setShowModal(true);
                                                     }}
                                                 >
                                                     <FaTrashAlt className="icon-delete" />
@@ -86,6 +96,48 @@ function ServiceList({ services, onServiceUpdate, onServiceEdit, userType, onCon
                     </div>
                 ))}
             </div>
+
+            {/* Modal */}
+            {showModal && (
+                <div
+                    className="modal fade show d-block"
+                    tabIndex="-1"
+                    role="dialog"
+                    style={{ backgroundColor: 'rgba(0, 0, 0, 0.5)' }}
+                >
+                    <div className="modal-dialog modal-dialog-centered" role="document">
+                        <div className="modal-content">
+                            <div className="modal-header">
+                                <h5 className="modal-title">Confirmar Eliminación</h5>
+                                <button
+                                    type="button"
+                                    className="btn-close"
+                                    onClick={() => setShowModal(false)}
+                                ></button>
+                            </div>
+                            <div className="modal-body">
+                                <p>¿Estás seguro de que deseas eliminar este servicio?</p>
+                            </div>
+                            <div className="modal-footer">
+                                <button
+                                    type="button"
+                                    className="btn btn-secondary"
+                                    onClick={() => setShowModal(false)}
+                                >
+                                    Cancelar
+                                </button>
+                                <button
+                                    type="button"
+                                    className="btn btn-warning"
+                                    onClick={handleDelete}
+                                >
+                                    Confirmar
+                                </button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            )}
         </div>
     );
 }
